@simaishi 2017-03-20
@kbrock Please resolve conflict and create an Euwe-specific PR (referencing this one) or suggest other PRs to backport.

```diff
$ git diff
diff --cc app/models/miq_report/generator.rb
index b118111,f340ead..0000000
--- a/app/models/miq_report/generator.rb
+++ b/app/models/miq_report/generator.rb
@@@ -85,7 -97,11 +85,15 @@@ module MiqReport::Generato
      if klass.nil?
        klass = db_class
        result = {}
++<<<<<<< HEAD
 +      cols.each { |c| result.merge!(c.to_sym => {}) if klass.virtual_attribute?(c) || klass == LiveMetric } if cols
++=======
+       if cols && klass.respond_to?(:virtual_attribute?)
+         cols.each do |c|
+           result[c.to_sym] = {} if klass.virtual_attribute?(c) && !klass.attribute_supported_by_sql?(c)
+         end
+       end
++>>>>>>> 832c5a3... Merge pull request #14285 from kbrock/widget_2
      end
  
      if includes.kind_of?(Hash)
@@@ -99,12 -115,16 +107,20 @@@
            assoc_klass = assoc_reflection.nil? ? nil : (assoc_reflection.options[:polymorphic] ? k : assoc_reflection.klass)
  
            if v.nil? || v["include"].blank?
 -            result[k] = {}
 -          elsif assoc_klass
 -            result[k] = include_as_hash(v["include"], assoc_klass)
 +            result.merge!(k => {})
 +          else
 +            result.merge!(k => get_include_for_find(v["include"], assoc_klass)) if assoc_klass
            end
  
++<<<<<<< HEAD
 +          v["columns"].each { |c| result[k].merge!(c.to_sym => {}) if assoc_klass.virtual_attribute?(c) } if assoc_klass && assoc_klass.respond_to?(:virtual_attribute?) && v["columns"]
++=======
+           if assoc_klass && assoc_klass.respond_to?(:virtual_attribute?) && v["columns"]
+             v["columns"].each do |c|
+               result[k][c.to_sym] = {} if assoc_klass.virtual_attribute?(c) && !assoc_klass.attribute_supported_by_sql?(c)
+             end
+           end
++>>>>>>> 832c5a3... Merge pull request #14285 from kbrock/widget_2
          end
        end
      elsif includes.kind_of?(Array)
```