@kbrock 2017-03-22
Thanks @simaishi - removing euwe label